refactor(services): Refactor structure, rename and move services
Ref #166
