--- app/services/api/base_service.rb
@@ -1,9 +1,0 @@
-module Api
-  class BaseService
-    def self.call(*args)
-      new(*args).call
-    end
-
-    def initialize(*args); end
-  end
-end
